+++ NugetDbTools/NugetDbPacker/Scripts/Publish-DbPackage.ps1
@@ -19,16 +19,19 @@ function Publish-DbPackage {
     $configPath = [IO.Path]::ChangeExtension($ProjectPath, '.nuget.config')
     $projFolder = Split-Path $ProjectPath -Resolve
     $nugetFolder = [IO.Path]::Combine($projFolder, 'NuGet')
+	$solutionFolder = Split-Path $SolutionPath
     if (Test-Path $configPath)
     {
         $settings = Import-NuGetSettings -NugetConfigPath $configPath -SolutionPath $SolutionPath
         $id = $settings.nugetSettings.Id
         $version = $settings.nugetSettings.version
         if (-not (Test-NuGetVersionExists -Id $id -Version $version)) {
-            $nugetPackage = [IO.Path]::Combine($nugetFolder, "$id.$version.nupkg")
+            $nugetPackage = [IO.Path]::Combine($solutionFolder, "$id.$version.nupkg")
             Initialize-DbPackage -ProjectPath $ProjectPath -SolutionPath $SolutionPath
-            Publish-NuGetPackage -PackagePath $nugetPackage
-            Remove-NugetFolder $nugetFolder
+            if ($env:USERNAME -EQ 'Builder') {
+				Publish-NuGetPackage -PackagePath $nugetPackage
+				Remove-NugetFolder $nugetFolder
+			}
         }
     }
 }
