+++ NugetDbTools/NugetDbPackerDb/PowerShell/NuGetDbPacker.psm1
@@ -1,5 +1,5 @@
 ﻿if (-not (Get-Module NugetSharedPacker -All)) {
-	Import-Module "$PSScriptRoot\NugetSharedPacker.psd1"
+	Import-Module "$PSScriptRoot\NugetSharedPacker.psd1" -Global
 }
 
 function Add-DbFileNode ($parentNode) {
@@ -35,8 +35,8 @@ function Enable-CLR{
 	if(-not (Get-InstalledModule SqlServer)) {
 		Install-Module SqlServer
 	}
-	Import-Module SqlServer
-	[xml]$doc = gc $ProfilePath
+	Import-Module SqlServer -Global -DisableNameChecking
+	[xml]$doc = Get-Content $ProfilePath
 	[string]$connectionString = $doc.Project.PropertyGroup.TargetConnectionString
 	$query = @'
 sp_configure 'show advanced options', 1;
@@ -90,13 +90,13 @@ function Find-PublishProfilePath {
 	$path = [IO.Path]::ChangeExtension($ProjectPath, '.publish.xml')
 	
 	$overrides = @($Override, $Host.Name, $hostType, (Get-Branch))
-	$overrides | ? {
+	$overrides | Where-Object {
 		-not [string]::IsNullOrEmpty($_) 
-	} | % {
+	} | ForEach-Object {
 		[IO.Path]::ChangeExtension($ProjectPath, ".$_.publish.xml")
-	} | ? {
+	} | Where-Object {
 		Test-Path $_
-	} | select -First 1 | % {
+	} | Select-Object -First 1 | ForEach-Object {
 		$path = $_
 	}
 	$path
@@ -104,22 +104,22 @@ function Find-PublishProfilePath {
 
 function Find-SqlPackagePath {
 	[IO.FileInfo]$info
-	ls "$env:ProgramFiles*\Microsoft Visual Studio\*\*\Common7\IDE\Extensions\Microsoft\SQLDB\DAC\*\SqlPackage.exe" |
-		sort -Property FullName -Descending |
-		select -First 1 | % {
+	Get-ChildItem "$env:ProgramFiles*\Microsoft Visual Studio\*\*\Common7\IDE\Extensions\Microsoft\SQLDB\DAC\*\SqlPackage.exe" |
+		Sort-Object -Property FullName -Descending |
+		Select-Object -First 1 | ForEach-Object {
 			$info = $_
 		}
 	if ($info -eq $null)  {
-		ls "$env:ProgramFiles*\Microsoft SQL Server\*\DAC\bin\SqlPackage.exe" |
-			sort -Property FullName -Descending |
-			select -First 1 | % {
+		Get-ChildItem "$env:ProgramFiles*\Microsoft SQL Server\*\DAC\bin\SqlPackage.exe" |
+			Sort-Object -Property FullName -Descending |
+			Select-Object -First 1 | ForEach-Object {
 			$info = $_
 		}
 	}
 	if ($info -eq $null)  {
-		ls "$env:ProgramFiles*\Microsoft Visual Studio*\Common7\IDE\Extensions\Microsoft\SQLDB\DAC\*\SqlPackage.exe" |
-			sort -Property FullName -Descending |
-			select -First 1 | % {
+		Get-ChildItem "$env:ProgramFiles*\Microsoft Visual Studio*\Common7\IDE\Extensions\Microsoft\SQLDB\DAC\*\SqlPackage.exe" |
+			Sort-Object -Property FullName -Descending |
+			Select-Object -First 1 | ForEach-Object {
 			$info = $_
 		}
 	}
@@ -130,9 +130,61 @@ function Find-SqlPackagePath {
 	}
 }
 
+function Format-ProjectDatabaseParameters {
+	<#.Synopsis
+	Format the SqlCommand CLI parameters to publish the DB project dacpac
+	.DESCRIPTION
+    Formats the parameters to publish the dacpac using profile if available and override parameters.
+	.EXAMPLE
+	Publish-ProjectDatabase -PublishTemplate C:\VSTS\EcsShared\SupportRoles\EcsShared.SupportRoles.publish.xml
+	#>
+    [CmdletBinding()]
+    [OutputType([string])]
+	param
+    (
+        # The location of .dacpac file being published
+		[string]$DacpacPath,
+        # The location of the profile (.publish.xml file being) with deployment options
+        [string]$ProfilePath,
+		# Parameters overriding profile settings
+		# Format according to SqlPackage CLI https://docs.microsoft.com/en-us/sql/tools/sqlpackage?view=sql-server-2017
+		[string[]]$Parameters
+	)
+
+	if (-not $DacpacPath) {
+		throw 'No DacPac was specified'
+	}
+	if (-not (Test-Path $DacpacPath)) {
+		throw "The DacPac does not exist at $DacpacPath"
+	}
+	if ($Parameters) {
+		$params = [string]::Join(' ', $Parameters)
+	} else {
+		$params = ''
+	}
+	if ($ProfilePath) {
+		if (Test-Path $ProfilePath) {
+			[string]$db = "/pr:`"$ProfilePath`" $params"
+		} else {
+			throw "The Profile does not exist at $ProfilePath"
+		}
+	} else {
+		if (-not ($params.Contains('/p:CreateNewDatabase'))) {
+			$params += ' /p:CreateNewDatabase=True'
+		}
+		if (-not ($params.Contains('/tdn:') -or $params.Contains('/TargetDatabaseName:'))) {
+			$projectName = [IO.Path]::GetFileNameWithoutExtension($DacpacPath)
+			[string]$db = "/tdn:`"$projectName`" $params"
+		} else {
+			[string]$db = $params
+		}
+	}
+	$db
+}
+
 function Import-NuGetDb {
 	<#.Synopsis
-	Copy the build files to the NuGet content folder
+	Copy-Item the build files to the NuGet content folder
 	.DESCRIPTION
 	Copies the dacpac and CLR assembly files to the NuGet content folder
 	.EXAMPLE
@@ -156,7 +208,7 @@ function Import-NuGetDb {
 	[xml]$proj = Get-Content $ProjectPath
 	[string]$dacpac = Get-ProjectProperty -Proj $proj -Property DacApplicationName
 	if ($dacpac -eq '') {
-		$dacpac = ([string]($proj.Project.PropertyGroup.Name | ? { $_ -ne 'PropertyGroup'})).Trim()
+		$dacpac = ([string]($proj.Project.PropertyGroup.Name | Where-Object { $_ -ne 'PropertyGroup'})).Trim()
 	}
 	[string]$assembly = Get-ProjectProperty -Proj $proj -Property AssemblyName
 
@@ -164,15 +216,15 @@ function Import-NuGetDb {
 		Copy-Item "$ProjDbFolder\$dacpac.dacpac" $NugetDbFolder
 	}
 	Copy-Item "$ProjDbFolder\*.*" $NugetDbFolder
-	ls $ProjDbFolder -Directory | % {
+	Get-ChildItem $ProjDbFolder -Directory | ForEach-Object {
 		$dir = $_.Name
-		md "$NugetDbFolder\$dir"  | Out-Null
+		mkdir "$NugetDbFolder\$dir"  | Out-Null
 		if (Test-Path "$ProjDbFolder\$dir\$dacpac.dacpac") {
 			Copy-Item "$ProjDbFolder\$dir\$dacpac.dacpac" "$NugetDbFolder\$dir"
 		}
 		Copy-Item "$ProjDbFolder\$dir\*.dll" "$NugetDbFolder\$dir"
 	}
-	[xml]$spec = gc $NugetSpecPath
+	[xml]$spec = Get-Content $NugetSpecPath
 	Add-DbFileNode -parentNode $spec.package
 	Out-FormattedXml -Xml $spec -FilePath $NugetSpecPath
 }
@@ -199,16 +251,19 @@ function Publish-DbPackage {
     $configPath = [IO.Path]::ChangeExtension($ProjectPath, '.nuget.config')
     $projFolder = Split-Path $ProjectPath -Resolve
     $nugetFolder = [IO.Path]::Combine($projFolder, 'NuGet')
+	$solutionFolder = Split-Path $SolutionPath
     if (Test-Path $configPath)
     {
         $settings = Import-NuGetSettings -NugetConfigPath $configPath -SolutionPath $SolutionPath
         $id = $settings.nugetSettings.Id
         $version = $settings.nugetSettings.version
         if (-not (Test-NuGetVersionExists -Id $id -Version $version)) {
-            $nugetPackage = [IO.Path]::Combine($nugetFolder, "$id.$version.nupkg")
+            $nugetPackage = [IO.Path]::Combine($solutionFolder, "$id.$version.nupkg")
             Initialize-DbPackage -ProjectPath $ProjectPath -SolutionPath $SolutionPath
-            Publish-NuGetPackage -PackagePath $nugetPackage
-            Remove-NugetFolder $nugetFolder
+            if ($env:USERNAME -EQ 'Builder') {
+				Publish-NuGetPackage -PackagePath $nugetPackage
+				Remove-NugetFolder $nugetFolder
+			}
         }
     }
 }
@@ -227,21 +282,18 @@ function Publish-ProjectDatabase {
         # The location of .dacpac file being published
 		[string]$DacpacPath,
         # The location of the profile (.publish.xml file being) with deployment options
-        [string]$ProfilePath
+        [string]$ProfilePath,
+		# Parameters overriding profile settings
+		# Format according to SqlPackage CLI https://docs.microsoft.com/en-us/sql/tools/sqlpackage?view=sql-server-2017
+		[string[]]$Parameters
 	)
 	[string]$cmd = Find-SqlPackagePath
 	if ($cmd) {
 		try {
-	
-			if ($ProfilePath -and (Test-Path $ProfilePath)) {
-				[string]$db = "/pr:`"$ProfilePath`""
-			} else {
-				$projectName = [IO.Path]::GetFileNameWithoutExtension($DacpacPath)
-				[string]$db = "/tdn:`"$projectName`" /p:CreateNewDatabase=True"
-			}
+			$params = Format-ProjectDatabaseParameters -DacpacPath $DacpacPath -ProfilePath $ProfilePath -Parameters $Parameters
 	
 			Log "Publishing $DacpacPath using $ProfilePath"
-			Invoke-Trap -Command "& `"$cmd`" /a:Publish /sf:`"$DacpacPath`" $db" -Message "Deploying database failed" -Fatal
+			Invoke-Trap -Command "& `"$cmd`" /a:Publish /sf:`"$DacpacPath`" $params" -Message "Deploying database failed" -Fatal
 		} catch {
 			Log "SqlPackage.exe failed: $_" -Error
 			exit 1
@@ -270,10 +322,242 @@ function Publish-SolutionDbPackages {
 	)
     $solutionFolder = Split-Path -Path $SolutionPath
 
-    Get-SqlProjects -SolutionPath $SolutionPath | % {
+    Get-SqlProjects -SolutionPath $SolutionPath | ForEach-Object {
         [string]$projectPath = [IO.Path]::Combine($solutionFolder, $_.ProjectPath)
         Publish-DbPackage -ProjectPath $projectPath -SolutionPath $SolutionPath
     }
+}
+
+function Publish-SSASCubeDatabase {
+	param (
+		[string]$CubeFolder,
+		[string]$CubeName,
+		[string]$ConfigSharedFolder,
+		[string]$ConfigFolder,
+		[string]$DatabaseName,
+		[string]$DeploymentError
+	)
+	if (Test-Path "$ConfigFolder") {
+		Update-SSASCubeAsdatabaseFile `
+			-SSASCubeAsdatabasePath "$CubeFolder\$CubeName.asdatabase" `
+			-DeploymentTargetConfigPath "$ConfigFolder\$CubeName.deploymenttargets"
+
+		Update-SSASCubeDeploymentOptions `
+			-SSASCubeDeploymentOptionsPath "$CubeFolder\$CubeName.deploymentoptions" `
+			-DeploymentOptionsConfigPath "$ConfigSharedFolder\default.deploymentoptions"
+
+		Update-SSASCubeDeploymentTarget `
+			-SSASCubeDeploymentTargetsPath "$CubeFolder\$CubeName.deploymenttargets" `
+			-DeploymentTargetConfigPath "$ConfigFolder\$CubeName.deploymenttargets"
+
+		Update-SSASCubeDataSource `
+			-SSASCubeConfigSettingsPath "$CubeFolder\$CubeName.configsettings" `
+			-DataSourceConfigPath "$ConfigFolder\DataSources.configsettings"
+
+		[string]$TargetServerVersion = "2016"
+
+		if(Test-Path -Path "$ConfigFolder\AdditionalSettings.xml") {
+		    [xml]$SSASCubeAdditionalSettingsFile = Get-Content "$ConfigFolder\AdditionalSettings.xml"
+			if (-not $SSASCubeAdditionalSettingsFile.Settings.Server.Version) {
+				Write-Host "SSAS Target server version variable not defined, defaulting to $TargetServerVersion" -fore red
+			} else {
+				$TargetServerVersion = $SSASCubeAdditionalSettingsFile.Settings.Server.Version;
+			}
+		} else {
+		    Write-Host "Additional settings file was not defined for this target deployment channel" -fore yellow
+		}
+		
+		Invoke-Trap -Command ".\Microsoft.AnalysisServices.Deployment.ps1 `"$CubeFolder\$CubeName.asdatabase`" `"$DatabaseName`" `"$TargetServerVersion`"" -Message $DeploymentError -Fatal
+	}
+}
+
+function Report-PublishProjectDatabase {
+	<#.Synopsis
+	Publish the DB project dacpac
+	.DESCRIPTION
+    Publishes the dacpac as specified by the publish template.
+	.EXAMPLE
+	Publish-ProjectDatabase -PublishTemplate C:\VSTS\EcsShared\SupportRoles\EcsShared.SupportRoles.publish.xml
+	#>
+    [CmdletBinding()]
+    param
+    (
+        # The location of .dacpac file being published
+		[string]$DacpacPath,
+        # The location of the profile (.publish.xml file being) with deployment options
+        [string]$ProfilePath,
+		# The destination of the deploy report
+		[string]$OutputPath,
+		# Parameters overriding profile settings
+		# Format according to SqlPackage CLI https://docs.microsoft.com/en-us/sql/tools/sqlpackage?view=sql-server-2017
+		[string[]]$Parameters
+	)
+	[string]$cmd = Find-SqlPackagePath
+	if ($cmd) {
+		try {
+			$params = Format-ProjectDatabaseParameters -DacpacPath $DacpacPath -ProfilePath $ProfilePath -Parameters $Parameters
+	
+			Log "Publishing $DacpacPath using $ProfilePath"
+			Invoke-Trap -Command "& `"$cmd`" /a:DeployReport /sf:`"$DacpacPath`" /op:`"$outputPath`" $params" -Message "Reporting the database deployment failed" -Fatal
+		} catch {
+			Log "SqlPackage.exe failed: $_" -Error
+			exit 1
+		}
+	} else {
+		Log "SqlPackage.exe could not be found" -E
+		exit 1
+	}}
+
+function Update-SSASCubeAsdatabaseFile {
+	param (
+		[string]$SSASCubeAsdatabasePath,
+		[string]$DeploymentTargetConfigPath
+	)
+	
+	try {
+		if ([string]::IsNullOrEmpty($SSASCubeAsdatabasePath)) {
+			Write-Host "$SSASCubeAsdatabasePath not found" -fore red
+			exit 1
+		}
+
+		if ([string]::IsNullOrEmpty($DeploymentTargetConfigPath)) {
+			Write-Host "$DeploymentTargetConfigPath not found" -fore red
+			exit 1
+		}
+
+		$SSASCubeAsdatabaseFileName = [IO.Path]::GetFileName($SSASCubeAsdatabasePath)
+		Write-Host "Updating $SSASCubeAsdatabaseFileName..."
+
+		[xml]$SSASCubeAsdatabaseFile = Get-Content "$SSASCubeAsdatabasePath"
+		[xml]$deploymentTargetConfigFile = Get-Content "$DeploymentTargetConfigPath"
+
+		$SSASCubeAsdatabaseFile.Database.ID = $deploymentTargetConfigFile.DeploymentTarget.Database
+		$SSASCubeAsdatabaseFile.Database.Name = $deploymentTargetConfigFile.DeploymentTarget.Database
+
+		$SSASCubeAsdatabaseFile.Save("$SSASCubeAsdatabasePath")
+
+		Write-Host "$SSASCubeAsdatabaseFileName datasource updated..."
+	} catch {
+		Log "Update-SSASCubeAsdatabaseFile failed: $_" -Error
+	}
+}
+
+function Update-SSASCubeDataSource {
+	param (
+		[string]$SSASCubeConfigSettingsPath,
+		[string]$DataSourceConfigPath
+	)
+
+	try {
+		if ([string]::IsNullOrEmpty($SSASCubeConfigSettingsPath)) {
+			Write-Host "$SSASCubeConfigSettingsPath not found" -fore red
+			exit 1
+		}
+
+		if ([string]::IsNullOrEmpty($DataSourceConfigPath)) {
+			Write-Host "$DataSourceConfigPath not found" -fore red
+			exit 1
+		}
+
+		$SSASCubeConfigSettingsFileName = [IO.Path]::GetFileName($SSASCubeConfigSettingsPath)
+		Write-Host "Updating $SSASCubeConfigSettingsFileName..."
+		
+		[xml]$SSASCubeConfigSettingsFile = Get-Content "$SSASCubeConfigSettingsPath"
+		[xml]$dataSourceConfigFile = Get-Content "$DataSourceConfigPath"
+
+		$configSettingsDataSources = $SSASCubeConfigSettingsFile.ConfigurationSettings.Database.DataSources.DataSource
+		
+		foreach ($dataSource in $dataSourceConfigFile.DataSources.DataSource) {
+			$configSettingDataSource = $configSettingsDataSources | Where-Object { $_.ID -eq $dataSource.ID }
+			$configSettingDataSource.ConnectionString = $dataSource.ConnectionString
+		}
+
+		$SSASCubeConfigSettingsFile.Save("$SSASCubeConfigSettingsPath")
+
+		Write-Host "$SSASCubeConfigSettingsFileName datasource updated..."
+	}
+	catch {
+		Log "Update-SSASCubeDetails failed: $_" -Error
+	}
+}
+
+function Update-SSASCubeDeploymentOptions {
+	param (
+		[string]$SSASCubeDeploymentOptionsPath,
+		[string]$DeploymentOptionsConfigPath
+	)
+
+	try {
+		if ([string]::IsNullOrEmpty($SSASCubeDeploymentOptionsPath)) {
+			Write-Host "$SSASCubeDeploymentOptionsPath not found" -fore red
+			exit 1
+		}
+
+		if ([string]::IsNullOrEmpty($DeploymentOptionsConfigPath)) {
+			Write-Host "$DeploymentOptionsConfigPath not found" -fore red
+			exit 1
+		}
+
+		$SSASdeploymentOptionsFileName = [IO.Path]::GetFileName($SSASCubeDeploymentOptionsPath)
+		Write-Host "Updating $SSASdeploymentOptionsFileName..."
+
+		[xml]$SSASCubeDeploymentOptionsFile = Get-Content "$SSASCubeDeploymentOptionsPath"
+		[xml]$deploymentOptionsConfigFile = Get-Content "$DeploymentOptionsConfigPath"
+		
+		$SSASCubeDeploymentOptions = $SSASCubeDeploymentOptionsFile.DeploymentOptions.ChildNodes
+		$deploymentOptions = $deploymentOptionsConfigFile.DeploymentOptions.ChildNodes
+
+		foreach ($deploymentOption in $deploymentOptions) {
+			$SSASCubeDeploymentOption = $SSASCubeDeploymentOptions | Where-Object { $_.Name -eq $deploymentOption.Name }
+
+			if ($null -ne $SSASCubeDeploymentOption) {
+				$SSASCubeDeploymentOption.InnerText = $deploymentOption.InnerText
+			}
+		}
+
+		$SSASCubeDeploymentOptionsFile.Save("$SSASCubeDeploymentOptionsPath")
+
+		Write-Host "$SSASdeploymentOptionsFileName datasource updated..."
+	}
+	catch {
+		Log "Update-SSASCubeDeploymentOptions failed: $_" -Error
+	}
+}
+
+function Update-SSASCubeDeploymentTarget {
+	param (
+		[string]$SSASCubeDeploymentTargetsPath,
+		[string]$DeploymentTargetConfigPath
+	)
+
+	try {
+		if ([string]::IsNullOrEmpty($SSASCubeDeploymentTargetsPath)) {
+			Write-Host "$SSASCubeDeploymentTargetsPath not found" -fore red
+			exit 1
+		}
+
+		if ([string]::IsNullOrEmpty($DeploymentTargetConfigPath)) {
+			Write-Host "$DeploymentTargetConfigPath not found" -fore red
+			exit 1
+		}
+
+		$SSASdeploymentTargetsFileName = [IO.Path]::GetFileName($SSASCubeDeploymentTargetsPath)
+		Write-Host "Updating $SSASdeploymentTargetsFileName..."
+
+		[xml]$SSASCubeDeploymentTargetsFile = Get-Content "$SSASCubeDeploymentTargetsPath"
+		[xml]$deploymentTargetConfigFile = Get-Content "$DeploymentTargetConfigPath"
+
+		$SSASCubeDeploymentTargetsFile.DeploymentTarget.Database = $deploymentTargetConfigFile.DeploymentTarget.Database
+		$SSASCubeDeploymentTargetsFile.DeploymentTarget.Server = $deploymentTargetConfigFile.DeploymentTarget.Server
+		$SSASCubeDeploymentTargetsFile.DeploymentTarget.ConnectionString = $deploymentTargetConfigFile.DeploymentTarget.ConnectionString
+
+		$SSASCubeDeploymentTargetsFile.Save("$SSASCubeDeploymentTargetsPath")
+
+		Write-Host "$SSASdeploymentTargetsFileName datasource updated..."
+	}
+	catch {
+		Log "Update-SSASCubeDeploymentTarget failed: $_" -Error
+	}
 }
 
 function Initialize-DbPackage
@@ -295,15 +579,17 @@ function Initialize-DbPackage
 		# The solution file
 		[string]$SolutionPath
 	)
+	$solutionFolder = Split-Path -LiteralPath $SolutionPath
 	$projectFolder = Split-Path -LiteralPath $ProjectPath -Resolve
-	$nugetPath = Join-Path -Path $projectFolder -ChildPath 'Nuget'
+	$nugetFolder = Join-Path -Path $projectFolder -ChildPath 'Nuget'
 	$configPath = [IO.Path]::ChangeExtension($ProjectPath, '.nuget.config')
 	$nugetSettings = Import-NuGetSettings -NugetConfigPath $configPath -SolutionPath $SolutionPath
+	$nuspecPath = "$nugetFolder\Package.nuspec"
 
 	Initialize-Package -ProjectPath $ProjectPath -NugetSettings $nugetSettings -SolutionPath $SolutionPath
-	mkdir "$nugetPath\content\Databases" | Out-Null
-	Import-NuGetDb -ProjectPath $ProjectPath -ProjDbFolder "$projectFolder\Databases" -NugetDbFolder "$nugetPath\content\Databases" -NugetSpecPath "$nugetPath\Package.nuspec"
-	Compress-Package -NugetPath $nugetPath
+	mkdir "$nugetFolder\content\Databases" | Out-Null
+	Import-NuGetDb -ProjectPath $ProjectPath -ProjDbFolder "$projectFolder\Databases" -NugetDbFolder "$nugetFolder\content\Databases" -NugetSpecPath "$nugetFolder\Package.nuspec"
+	Compress-Package -NuspecPath $nuspecPath -NugetFolder $nugetFolder -PackageFolder $solutionFolder
 }
